requirements.txt
streamlit==1.28.1
google-generativeai>=0.3.0
python-dotenv==1.0.0
PyMuPDF>=1.23.0
python-docx==0.8.11
pandas>=2.1.0
openpyxl>=3.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.0
deap>=1.4.0
matplotlib>=3.8.0
plotly>=5.17.0
pydantic>=2.0.0
